requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain-google-genai
chromadb
langchain-chroma
python-dotenv
rapidfuzz
pydantic